move config to temporalworker
